requirements.txt
pillow
pygame
aiohttp
requests
python-bidi
flask